Refactoring Zed
