+++ src/Pyz/Zed/AiAnythingVectorDb/AiAnythingVectorDbDependencyProvider.php
@@ -1,26 +1,30 @@
 <?php
 
-namespace Pyz\Zed\ProductVectorDbStorage;
+namespace Pyz\Zed\AiAnythingVectorDb;
+
 use Spryker\Zed\Kernel\AbstractBundleDependencyProvider;
 use Spryker\Zed\Kernel\Container;
 
-class ProductVectorDbStorageDependencyProvider extends AbstractBundleDependencyProvider
+class AiAnythingVectorDbDependencyProvider extends AbstractBundleDependencyProvider
 {
-
+    /**
+     * @var string
+     */
     public const PINECONE_CLIENT = 'PINECONE_CLIENT';
+
     public function provideBusinessLayerDependencies(Container $container): Container
-    {   
+    {
         $container = parent::provideBusinessLayerDependencies($container);
 
-        $container = $this->addPineconeClient($container);
+        $container = $this->addVectorDbClient($container);
 
         return $container;
     }
 
-    public function addPineconeClient($container): Container {
-        
+    public function addVectorDbClient($container): Container
+    {
         $container->set(static::PINECONE_CLIENT, function (Container $container) {
-            return $container->getLocator()->pinecone()->client();
+            return $container->getLocator()->aiAnythingVectorDb()->client();
         });
 
         return $container;
