+++ src/Pyz/Zed/AiAnythingVectorDb/Business/AiAnythingVectorDbFacadeInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pyz\Zed\AiAnythingVectorDb\Business;
+
+interface AiAnythingVectorDbFacadeInterface
+{
+    public function writeProductDataToVectorDb(array $transfers):void;
+}
