+++ src/Pyz/Zed/AiAnythingVectorDb/Persistence/AiAnythingVectorDbRepositoryInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Pyz\Zed\AiAnythingVectorDb\Persistence;
+
+interface AiAnythingVectorDbRepositoryInterface
+{
+    /**
+     * @param string $id
+     *
+     * @return array
+     */
+    public function queryProductData(string $id): array;
+}
