--- src/Pyz/Zed/ProductVectorDbStorage/Business/ProductVectorDbStorageFacadeInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Pyz\Zed\ProductVectorDbStorage\Business;
-
-interface ProductVectorDbStorageFacadeInterface
-{
-    public function writeProductDataToVectorDb(array $transfers):void;
-}
