--- src/Pyz/Zed/ProductVectorDbStorage/Persistence/ProductVectorDbStorageRepositoryInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Pyz\Zed\ProductVectorDbStorage\Persistence;
-
-interface ProductVectorDbStorageRepositoryInterface
-{
-
-}
